Autofocus URL textbox, rename link toolbars
